mock: update to lexicon
