+++ packages/lexicon/package.json
@@ -3,5 +3,12 @@
   "version": "1.0.0",
   "description": "Central repo of names and type definitions",
   "author": "Open Function Group <admin@openfn.org>",
-  "license": "ISC"
+  "license": "ISC",
+  "exports": {
+    "lightning": {
+      "import": {
+        "types": "./lightning.d/.ts"
+      }
+    }
+  }
 }
