+++ packages/lightning-mock/package.json
@@ -18,6 +18,7 @@
   "dependencies": {
     "@koa/router": "^12.0.0",
     "@openfn/engine-multi": "workspace:*",
+    "@openfn/lexicon": "workspace:^",
     "@openfn/logger": "workspace:*",
     "@openfn/runtime": "workspace:*",
     "@types/koa-logger": "^3.1.2",
