+++ packages/lightning-mock/tsconfig.json
@@ -1,6 +1,6 @@
 {
   "extends": "../../tsconfig.common",
-  "include": ["src/**/*.ts"],
+  "include": ["src/**/*.ts", "test/**/*.ts"],
   "compilerOptions": {
     "module": "ESNext"
   }
